+++ src-tauri/src/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod autorun;
 pub mod dpapi;
+pub mod fs;